// packages/coze-js/src/ws-tools/chat/types.ts
export interface AudioCaptureConfig {
  /**
   * en: Whether to enable noise suppression
   * zh: 是否启用降噪
   */
  noiseSuppression?: boolean;
  /**
   * en: Whether to enable echo cancellation
   * zh: 是否启用回声消除
   */
  echoCancellation?: boolean;
  /**
   * en: Whether to enable automatic gain control
   * zh: 是否启用自动增益控制
   */
  autoGainControl?: boolean;
}
export interface AIDenoisingConfig {
  /**
   * en: AI denoiser mode
   * - `NSNG`: Enable non-stationary noise suppression.
   * - `STATIONARY_NS`: Enable stationary noise suppression.
   * zh: AI 降噪
   * - `NSNG`: 启用非平稳噪声抑制
   * - `STATIONARY_NS`: 启用平稳噪声抑制
   */
  mode?: 'NSNG' | 'STATIONARY_NS';
  /**
   * en: AI denoiser level
   * - `SOFT`: (Recommended) Soft mode
   * - `AGGRESSIVE`: Aggressive mode
   * zh: AI 降噪等级
   * - `SOFT`: （推荐）舒缓降噪
   * - `AGGRESSIVE`: 激进降噪。将降噪强度提高到激进降噪会增大损伤人声的概率。
   */
  level?: 'SOFT' | 'AGGRESSIVE';
  /**
   * en: The path to the AI denoiser assets, default is `/external`
   * zh: AI 降噪 wasm 文件路径，默认是 `/external`
   */
  assetsPath?: string;
}

export interface WavRecordConfig {
  /**
   * en: Whether to enable source record
   * zh: 是否启用源音频录制
   */
  enableSourceRecord: boolean;
  /**
   * en: Whether to enable denoise record
   * zh: 是否启用降噪后音频录制
   */
  enableDenoiseRecord: boolean;
}